use const static instead of read_only wherever possible
read_only doesn't give type info, or info to the optimizer (I think)
about the memory being read only. const does both. read only should only
be used when the memory is a gravestone value to something that's
normally mutable, to guarantee that the gravestone never changes, while
not giving it the const type contamination thingy